added marker insert feature with doubt tap, added custom view for markers, updated add marker dialog ui.

+++ app/src/main/java/com/cybershark/drawingsapp/data/models/MarkerEntity.kt
@@ -12,6 +12,6 @@ data class MarkerEntity(
     val description: String,
     val remarks: String,
     val assignee: String,
-    val markerPositionX: Long,
-    val markerPositionY: Long
+    val markerPositionX: Float,
+    val markerPositionY: Float
 )
